requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
dj-email-url==1.0.2
dj-rest-auth==2.1.4
Django==3.2.2
django-allauth==0.44.0
django-cache-url==3.2.3
djangorestframework==3.12.4
environs==9.3.2
idna==2.10
marshmallow==3.12.1
oauthlib==3.1.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.17.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.4
